refactor(oneseo): GraduationStatus enum class를 oneseo.entity.type 패키지로 이동
