+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/response/ApplicationDto.java
@@ -2,11 +2,10 @@ package team.themoment.hellogsmv3.domain.application.dto.response;
 
 import lombok.Builder;
 import team.themoment.hellogsmv3.domain.application.type.EvaluationStatus;
-import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.GraduationStatus;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 
 import java.math.BigDecimal;
-import java.util.UUID;
 
 @Builder
 public record ApplicationDto(
