+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/response/GedAdmissionInfoResDto.java
@@ -3,7 +3,7 @@ package team.themoment.hellogsmv3.domain.application.dto.response;
 import lombok.Builder;
 import team.themoment.hellogsmv3.domain.member.entity.type.Gender;
 import team.themoment.hellogsmv3.domain.application.type.DesiredMajors;
-import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.GraduationStatus;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 
 import java.time.LocalDate;
