+++ src/main/java/team/themoment/hellogsmv3/domain/application/service/QueryAllApplicationService.java
@@ -14,7 +14,7 @@ import team.themoment.hellogsmv3.domain.application.entity.CandidatePersonalInfo
 import team.themoment.hellogsmv3.domain.application.entity.abs.AbstractApplication;
 import team.themoment.hellogsmv3.domain.application.entity.abs.AbstractPersonalInformation;
 import team.themoment.hellogsmv3.domain.application.repo.ApplicationRepository;
-import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.GraduationStatus;
 
 import java.util.ArrayList;
 import java.util.List;
