+++ src/main/java/team/themoment/hellogsmv3/domain/application/service/QueryApplicationByIdService.java
@@ -8,7 +8,7 @@ import team.themoment.hellogsmv3.domain.application.dto.response.*;
 import team.themoment.hellogsmv3.domain.application.entity.*;
 import team.themoment.hellogsmv3.domain.application.entity.abs.AbstractApplication;
 import team.themoment.hellogsmv3.domain.application.repo.ApplicationRepository;
-import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.GraduationStatus;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 import team.themoment.hellogsmv3.global.exception.error.ExpectedException;
