+++ src/main/java/team/themoment/hellogsmv3/domain/application/service/SearchApplicationService.java
@@ -15,7 +15,7 @@ import team.themoment.hellogsmv3.domain.application.entity.CandidatePersonalInfo
 import team.themoment.hellogsmv3.domain.application.entity.abs.AbstractApplication;
 import team.themoment.hellogsmv3.domain.application.entity.abs.AbstractPersonalInformation;
 import team.themoment.hellogsmv3.domain.application.repo.ApplicationRepository;
-import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
+import team.themoment.hellogsmv3.domain.oneseo.entity.type.GraduationStatus;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
 import team.themoment.hellogsmv3.domain.application.type.SearchTag;
