+++ src/main/java/team/themoment/hellogsmv3/domain/oneseo/entity/type/GraduationStatus.java
@@ -1,4 +1,4 @@
-package team.themoment.hellogsmv3.domain.application.type;
+package team.themoment.hellogsmv3.domain.oneseo.entity.type;
 
 public enum GraduationStatus {
     CANDIDATE,
